refactoring tail settings component to use page util service
